requirements.txt
blinker==1.8.0
bson==0.5.10
click==8.1.7
DateTime==5.5
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
pip==24.0
pymongo==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.5.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.2
zope.interface==6.3
